chore: remove unused and rename mod
