+++ foyer/src/lib.rs
@@ -15,16 +15,6 @@
 #![feature(trait_alias)]
 #![feature(pattern)]
 
-pub trait Weight {
-    fn weight(&self) -> usize;
-}
-
-impl Weight for Vec<u8> {
-    fn weight(&self) -> usize {
-        self.len()
-    }
-}
-
 pub use foyer_common as common;
 pub use foyer_intrusive as intrusive;
 pub use foyer_storage as storage;
